Update Podman shell script
- Expanded podman-build shell script. It can now be used for
creating images and containers. It uses flags to make distinction
between the two operations. When creating a container the argument
after the flag must be the image id, which makes sure that both the
image and the container include in their names the same application
version and the same short commit hash.