Fix token not passed to the back-end
- Updated TranslationApiService to retrieve the token via
AuthStorageService from the browser session. Previously, the
code tried to access the token in the browser local storage, but the
token was not stored there, so null was passed instead of the token, so
even through the user was logged it, the application was acting as if
the user was not logged in.