Update the translation warning message
- Created one warning message for two different cases: when the character
limit is trespassed and when body of the scraped page is null due to
http 403 response.
In the future such warnings will not be shown at all to the user. There
will always be a translated page returned. If there is 403 returned the
application will try to scrape a different page.